Move `MeanSolarDistance` to `libOPHD`
