+++ appOPHD/MapObjects/Structures/SolarPanelArray.cpp
@@ -1,6 +1,6 @@
 #include "SolarPanelArray.h"
 
-#include "../../MeanSolarDistance.h"
+#include <libOPHD/MeanSolarDistance.h>
 
 
 namespace
