+++ libOPHD/MeanSolarDistance.cpp
