+++ libOPHD/MeanSolarDistance.h
